src/main/java/xyz/itbs/recipes/converters/NotesCommandToNotes.java
package xyz.itbs.recipes.converters;

import lombok.Synchronized;
import org.springframework.core.convert.converter.Converter;
import org.springframework.lang.Nullable;
import org.springframework.stereotype.Component;
import xyz.itbs.recipes.commands.NotesCommand;
import xyz.itbs.recipes.domain.Notes;

@Component
public class NotesCommandToNotes implements Converter<NotesCommand, Notes> {

    @Synchronized
    @Nullable
    @Override
    public Notes convert(NotesCommand source) {
        if(source == null) {
            return null;
        }
        final Notes notes = Notes.builder()
                .id(source.getId())
                .notesBody(source.getNotesBody())
                .build();
        return notes;
    }
}
